# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NiazFinder API",
    "version": "0.2.x-unstable",
    "description": "Generated skeleton of the Next.js App Router API surface. Methods and auth requirements are statically detected; only operations with explicit request/response schemas are stable. Regenerate with `npm run docs:openapi`.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/sasanappstore2/NiazFinder-main/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev (`npm run dev`)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "User token from phone-OTP login (AuthToken)."
      },
      "internalSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-secret",
        "description": "INTERNAL_API_SECRET for worker/cron routes. Constant-time compare, fail-closed."
      }
    }
  },
  "paths": {
...
    "/api/requests/{id}": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "GET /api/requests/[id]",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "requests"
        ],
        "summary": "PUT /api/requests/[id]",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "requests"
        ],
        "summary": "DELETE /api/requests/[id]",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/super-admin/businesses/{id}/moderate": {
      "post": {
        "tags": [
          "super-admin"
        ],
        "summary": "POST /api/super-admin/businesses/[id]/moderate",
        "description": "No server-side auth call detected statically — verify before treating as public.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      }
    },
...
  }
}